Added the ArrowTube and ArrowBullet which firing with arrow. Added the Fire and FireTube which using ParticleEffect and ParticleEmitter Added the variations of Bird - FastBird and SlowBird
in Bird.java
Modified constructor of Bird for
Added the variable gold - the cost of Bird

 in Bullet.java
 added the logic of dead
 and the if-statement when it stop check the Bullet as isDead
 added the sound by default

 in Tube.java
 added the firerate, value of Tube

 in PlayState
added the logic of money
added the logic of firing with Fire.java
+ minor fixes

Also added the functionality that didn't allow to closed Tube to fire. And make the Arrow to move a little bit smoothly.